added enum for sections' titles

+++ src/test/kotlin/com/github/darderion/mundaneassignmentpolice/checker/RulesTests.kt
@@ -50,13 +50,13 @@ class RulesTests: StringSpec({
 		RULE_SYMBOLS_IN_SECTION_NAMES.process(PDFBox().getPDF(filePathSymbolsInSectionNames)).count() shouldBeExactly 4
 	}
 	"Section rules should detect sections whose size exceeds specified limit" {
-		(introductionSizeRule as SectionSizeRule).percentageLimit shouldBeExactly 20
+		(introductionSizeRule as SectionSizeRule).percentageLimit.toInt() shouldBeExactly 20
 		introductionSizeRule.process(PDFBox().getPDF(filePathIntroductionSize)).count() shouldBeExactly 1
 
 		(problemStatementSizeRule as SectionSizeRule).pageLimit shouldBeExactly 1
 		problemStatementSizeRule.process(PDFBox().getPDF(filePathProblemStatementSize)).count() shouldBeExactly 1
 
-		(overviewSizeRule as SectionSizeRule).percentageLimit shouldBeExactly 50
+		(overviewSizeRule as SectionSizeRule).percentageLimit.toInt() shouldBeExactly 50
 		overviewSizeRule.process(PDFBox().getPDF(filePathOverviewSize)).count() shouldBeExactly 1
 
 		(conclusionSizeRule as SectionSizeRule).pageLimit shouldBeExactly 2
